requirements.txt
docopt>=0.6.2
GitPython>=3.0.5
numpy>=1.17.4